requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
pydantic==2.11.7
openai==1.101.0
faiss-cpu==1.11.0
sentence-transformers==5.1.0
pypdf==5.9.0
duckduckgo-search==8.1.1
gTTS==2.5.4
numpy==2.3.2
python-dotenv==1.1.1
